popup now appears in dashboard as requested and will only appear once in a user's session
